move db_operations tests
